Abstract out the SSH key fields
